+++ app/lib/ui/about/about_screen.dart
@@ -14,6 +14,9 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:ouds_core/components/common/ouds_icon_status.dart';
 import 'package:ouds_core/components/tag/ouds_tag.dart';
+import 'package:ouds_core/components/typography/ouds_annotated_text.dart';
+import 'package:ouds_core/components/typography/ouds_body_text.dart';
+import 'package:ouds_core/components/typography/ouds_heading_text.dart';
 import 'package:ouds_flutter_demo/l10n/app_localizations.dart';
 import 'package:ouds_flutter_demo/ui//utilities/settings_helper.dart';
 import 'package:ouds_flutter_demo/ui/about/detail/about_file_screen.dart';
@@ -62,20 +65,24 @@ class _AboutScreenState extends State<AboutScreen> {
     setState(() {});
   }
 
-  Widget _buildLabeledTag(OudsThemeContract currentTheme, String text, String version){
-    return  Column(
+  Widget _buildLabeledTag(
+    OudsThemeContract currentTheme,
+    String text,
+    String version,
+  ) {
+    return Column(
       children: [
         Row(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            Expanded(
-              child: Text(
-                text,
-                style: currentTheme.typographyTokens.typeBodyDefaultMedium(context),
-              ),
-            ),
+            Expanded(child: OudsBodyText(text: text)),
             SizedBox(width: currentTheme.spaceScheme(context).rowGapSmall),
-            OudsTag.text(label: version, status: Info(), appearance: OudsTagAppearance.muted, size: OudsTagSize.small)
+            OudsTag.text(
+              label: version,
+              status: Info(),
+              appearance: OudsTagAppearance.muted,
+              size: OudsTagSize.small,
+            ),
           ],
         ),
         SizedBox(height: currentTheme.spaceScheme(context).rowGapMedium),
@@ -95,117 +102,132 @@ class _AboutScreenState extends State<AboutScreen> {
       l10n.app_about_details_androidSystemVersion: Environment.androidSystem,
       l10n.app_about_details_themeOrangeCoreVersion: Environment.orangeCore,
       l10n.app_about_details_themeOrangeBrandVersion: Environment.orangeBrand,
-      l10n.app_about_details_themeOrangeCompactBrandVersion: Environment.orangeCompactBrand,
+      l10n.app_about_details_themeOrangeCompactBrandVersion:
+          Environment.orangeCompactBrand,
       l10n.app_about_details_themeCoreVersion: Environment.oudsCore,
       l10n.app_about_details_themeSoshCoreVersion: Environment.soshCore,
       l10n.app_about_details_themeSoshBrandVersion: Environment.soshBrand,
-      l10n.app_about_details_themeWireframeCoreVersion: Environment.wireframeCore,
-      l10n.app_about_details_themeWireframeBrandVersion: Environment.wireframeBrand,
+      l10n.app_about_details_themeWireframeCoreVersion:
+          Environment.wireframeCore,
+      l10n.app_about_details_themeWireframeBrandVersion:
+          Environment.wireframeBrand,
     };
     return ListView(
+      children: [
+        Column(
           children: [
-            Column(
-              children: [
-                Padding(
-                  padding: EdgeInsetsDirectional.all(
-                      currentTheme.spaceScheme(context).scaledMedium,
+            Padding(
+              padding: EdgeInsetsDirectional.all(
+                currentTheme.spaceScheme(context).scaledMedium,
+              ),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: [
+                  OudsHeadingText(
+                    text: context.l10n.app_about_name_label,
+                    size: OudsHeadingTextSize.xLarge,
                   ),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: [
-                      Text(
-                        context.l10n.app_about_name_label,
-                        style: currentTheme.typographyTokens.typeHeadingXLarge(context),
-                      ),
-                      SizedBox(height: currentTheme.spaceScheme(context).rowGapSmall),
-                      Text(
+                  SizedBox(
+                    height: currentTheme.spaceScheme(context).rowGapSmall,
+                  ),
+                  OudsBodyText(
+                    text:
                         '${currentEnvironment.name[0].toUpperCase()}${currentEnvironment.name.substring(1)} version ${_packageInfo.version} (${_packageInfo.buildNumber}) - Flutter',
-                        style: currentTheme.typographyTokens.typeBodyStrongLarge(context),
-                      ),
-                      SizedBox(height: currentTheme.spaceScheme(context).rowGapLarge),
-                      ...versionMap.entries.map((e) => _buildLabeledTag(currentTheme, e.key, e.value)),
-                    ],
+                    size: OudsBodyTextSize.large,
+                    weight: OudsTextWeight.strong,
                   ),
-                ),
-                ListTile(
-                  title: Text(
-                    context.l10n.app_about_legalInformation_label,
-                    style: currentTheme.typographyTokens.typeBodyStrongLarge(context),
+                  SizedBox(
+                    height: currentTheme.spaceScheme(context).rowGapLarge,
                   ),
-                  onTap: () {
-                    Get.to(
-                      AboutFileScreen(
-                        title: context.l10n.app_about_legalInformation_label,
-                        fileMenuItem: 'assets/about_legal_information.md',
-                        darkModeEnabled: Theme.of(context).brightness == Brightness.light,
-                        previousPageTitle: context.l10n.app_bottomBar_about_label,
-                      ),
-                    );
-                  },
-                ),
-                ListTile(
-                  title: Text(
-                    context.l10n.app_about_privacyPolicy_label,
-                    style: currentTheme.typographyTokens.typeBodyStrongLarge(context),
+                  ...versionMap.entries.map(
+                    (e) => _buildLabeledTag(currentTheme, e.key, e.value),
                   ),
-                  onTap: () {
-                    Get.to(
-                      AboutFileScreen(
-                        title: context.l10n.app_about_privacyPolicy_label,
-                        fileMenuItem: 'assets/about_privacy_policy.md',
-                        darkModeEnabled: Theme.of(context).brightness == Brightness.light,
-                        previousPageTitle: context.l10n.app_bottomBar_about_label,
-                      ),
-                    );
-                  },
-                ),
-                ListTile(
-                  title: Text(
-                    context.l10n.app_about_changelog_label,
-                    style: currentTheme.typographyTokens.typeBodyStrongLarge(context),
+                ],
+              ),
+            ),
+            ListTile(
+              title: OudsBodyText(
+                text: context.l10n.app_about_legalInformation_label,
+                size: OudsBodyTextSize.large,
+                weight: OudsTextWeight.strong,
+              ),
+              onTap: () {
+                Get.to(
+                  AboutFileScreen(
+                    title: context.l10n.app_about_legalInformation_label,
+                    fileMenuItem: 'assets/about_legal_information.md',
+                    darkModeEnabled:
+                        Theme.of(context).brightness == Brightness.light,
+                    previousPageTitle: context.l10n.app_bottomBar_about_label,
                   ),
-                  onTap: () {
-                    Get.to(
-                      AboutFileScreen(
-                        title: context.l10n.app_about_changelog_label,
-                        fileMenuItem: 'CHANGELOG.md',
-                        darkModeEnabled: Theme.of(context).brightness == Brightness.light,
-                        previousPageTitle: context.l10n.app_bottomBar_about_label,
-                      ),
-                    );
-                  },
-                ),
-                ListTile(
-                  title: Text(
-                    context.l10n.app_about_materialComponents_label,
-                    style: currentTheme.typographyTokens.typeBodyStrongLarge(context),
+                );
+              },
+            ),
+            ListTile(
+              title: OudsBodyText(
+                text: context.l10n.app_about_privacyPolicy_label,
+                size: OudsBodyTextSize.large,
+                weight: OudsTextWeight.strong,
+              ),
+              onTap: () {
+                Get.to(
+                  AboutFileScreen(
+                    title: context.l10n.app_about_privacyPolicy_label,
+                    fileMenuItem: 'assets/about_privacy_policy.md',
+                    darkModeEnabled:
+                        Theme.of(context).brightness == Brightness.light,
+                    previousPageTitle: context.l10n.app_bottomBar_about_label,
                   ),
-                  onTap: () {
-                    Get.to(
-                      MaterialComponentScreen(
-                        scaffoldKey: scaffoldKey,
-                        title: context.l10n.app_about_materialComponents_label,
-                        previousPageTitle: context.l10n.app_bottomBar_about_label,
-                      ),
-                    );
-                  },
-                ),
-                ListTile(
-                  title: Text(
-                    context.l10n.app_about_appSettings_label,
-                    style: currentTheme.typographyTokens.typeBodyDefaultLarge(context).copyWith(
-                      color: currentTheme.colorScheme(context).contentBrandPrimary,
-                    ),
+                );
+              },
+            ),
+            ListTile(
+              title: OudsBodyText(
+                text: context.l10n.app_about_changelog_label,
+                size: OudsBodyTextSize.large,
+                weight: OudsTextWeight.strong,
+              ),
+              onTap: () {
+                Get.to(
+                  AboutFileScreen(
+                    title: context.l10n.app_about_changelog_label,
+                    fileMenuItem: 'CHANGELOG.md',
+                    darkModeEnabled:
+                        Theme.of(context).brightness == Brightness.light,
+                    previousPageTitle: context.l10n.app_bottomBar_about_label,
                   ),
-                  onTap: () {
-                    Get.to(
-                      SettingsHelper.openAppropriateSettings(),
-                    );
-                  },
-                ),
-              ],
-            )
-          ] ,
+                );
+              },
+            ),
+            ListTile(
+              title: OudsBodyText(
+                text: context.l10n.app_about_materialComponents_label,
+                size: OudsBodyTextSize.large,
+                weight: OudsTextWeight.strong,
+              ),
+              onTap: () {
+                Get.to(
+                  MaterialComponentScreen(
+                    scaffoldKey: scaffoldKey,
+                    title: context.l10n.app_about_materialComponents_label,
+                    previousPageTitle: context.l10n.app_bottomBar_about_label,
+                  ),
+                );
+              },
+            ),
+            ListTile(
+              title: OudsBodyText(
+                text: context.l10n.app_about_appSettings_label,
+                size: OudsBodyTextSize.large,
+                weight: OudsTextWeight.strong,
+              ),
+              onTap: () {
+                Get.to(SettingsHelper.openAppropriateSettings());
+              },
+            ),
+          ],
+        ),
+      ],
     );
   }
 }