+++ app/lib/ui/tokens/color/color_tokens_model.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:ouds_flutter_demo/l10n/app_localizations.dart';
-import 'package:ouds_flutter_demo/ui/tokens/color/color_screen.dart';
+import 'package:ouds_flutter_demo/ui/utilities/colors_utils.dart';
 import 'package:ouds_theme_contract/ouds_theme_contract.dart';
 
 class ColorTokensModel {